implement user profile live update using friend status
