+++ front/app/src/js/components/pages/user_profile/UserProfileHeader.js
@@ -1,5 +1,7 @@
 import {Component} from '@components';
 import {userManagementClient} from '@utils/api';
+import {ErrorPage} from '@utils/ErrorPage.js';
+import {FriendsCache} from '@utils/cache';
 
 export class UserProfileHeader extends Component {
   constructor() {
@@ -70,7 +72,7 @@ export class UserProfileHeader extends Component {
                     <h1 class="mt-4 ${placeholderClass} col-12">_</h1>
                     <div class="d-flex">
                         <button class="btn btn-sm me-2 ${placeholderClass}">
-                            Follow
+                            Send Friend Request
                         </button>
                         <button class="btn btn-sm ${placeholderClass}">
                             Challenge
@@ -99,9 +101,7 @@ export class UserProfileHeader extends Component {
   renderProfileInteractionButtons() {
     return (`
       <div class="d-flex">
-          <button class="btn btn-sm btn-outline-success me-2">
-              Follow
-          </button>
+        ${this.#renderFriendButton()}
           <button class="btn btn-sm btn-outline-secondary">
               Challenge
           </button>
@@ -109,7 +109,9 @@ export class UserProfileHeader extends Component {
     `);
   }
 
-  loadUserProfile(username) {
+  loadUserProfile(username, userId) {
+    this.userId = userId;
+    this.username = username;
     this.headerContainer = this.querySelector('#header-container');
     if (this.headerContainer) {
       this.headerContainer.classList.remove('placeholder-glow');
@@ -120,6 +122,86 @@ export class UserProfileHeader extends Component {
           username, '/img/tdameros.jpg',
           username === userManagementClient.username,
       );
+      this.friendActionBtn = this.querySelector('#friend-action-btn');
+      super.addComponentEventListener(
+          this.friendActionBtn, 'click', this.#friendActionHandler,
+      );
+      super.addComponentEventListener(
+          document, FriendsCache.event, () => {
+            if (this.friendActionBtn) {
+              this.friendActionBtn.outerHTML = this.#renderFriendButton();
+              this.friendActionBtn = this.querySelector('#friend-action-btn');
+              super.addComponentEventListener(
+                  this.friendActionBtn, 'click', this.#friendActionHandler,
+              );
+            }
+          },
+      );
+    }
+  }
+
+  #renderFriendButton() {
+    const friend = FriendsCache.get(this.userId);
+    if (!friend) {
+      return (`
+        <button id="friend-action-btn" friend-action="send-request" class="btn btn-sm btn-success me-2">
+            Send Friend Request
+        </button>
+      `);
+    } else if (friend['status'] === 'pending') {
+      return (`
+        <button id="friend-action-btn" class="btn btn-sm btn-success me-2" disabled>
+            Friend Request Sent
+        </button>
+      `);
+    } else {
+      return (`
+        <button id="friend-action-btn" friend-action="remove-friend" class="btn btn-sm btn-danger me-2">
+          Remove Friend
+        </button>
+      `);
+    }
+  }
+
+  async #friendActionHandler() {
+    const action = this.friendActionBtn.getAttribute('friend-action');
+    if (action === 'send-request') {
+      await this.#addFriend();
+    } else if (action === 'remove-friend') {
+      await this.#removeFriend();
+    }
+  }
+
+  async #addFriend() {
+    try {
+      const {response} = await userManagementClient.sendFriendRequest(
+          this.userId,
+      );
+      if (response.ok) {
+        FriendsCache.set(this.userId, {
+          'id': this.userId,
+          'username': this.username,
+          'status': 'pending',
+          'connected_status': 'unknow',
+        });
+        document.querySelector('friends-component').updateFriends();
+      }
+    } catch {
+      ErrorPage.loadNetworkError();
+    }
+  }
+
+  async #removeFriend() {
+    try {
+      const {response} = await userManagementClient.removeFriend(
+          this.userId,
+      );
+      if (response.ok) {
+        FriendsCache.delete(this.userId);
+        document.querySelector('friends-component').updateFriends();
+      }
+    } catch {
+      ErrorPage.loadNetworkError();
     }
   }
 }
